+++ src/definitions.hpp
@@ -96,16 +96,8 @@ matmodules.cpp
     _sim = sim; \
     _name = name
 
-#define BIT0                0x01
-#define BIT1                0x02
-#define BIT2                0x04
-#define BIT3                0x08
-#define BIT4                0x10
-#define BIT5                0x20
-#define BIT6                0x40
-#define BIT7                0x80
-#define BIT_01              (BIT0+BIT1)
-#define BUS_SIZED           BIT0
-#define BUS_INITIALIZED     BIT_01
+#define BUS_SIZED           0x01
+#define BUS_GENERATED       0x03
+#define BUS_INITIALIZED     0X07
 
 #endif // DEFINITIONS_H